Break out deposit wallet UI API module in submodules
