+++ lib/ui/src/Cardano/Wallet/UI/Deposit/API.hs
@@ -1,15 +1,6 @@
-{-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DuplicateRecordFields #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE PolyKinds #-}
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -36,6 +27,9 @@ import Cardano.Wallet.UI.Common.Handlers.SSE
 import Cardano.Wallet.UI.Cookies
     ( CookieRequest
     )
+import Cardano.Wallet.UI.Deposit.API.Addresses.Transactions
+    ( TransactionHistoryParams
+    )
 import Control.Lens
     ( makePrisms
     )
@@ -56,19 +50,6 @@ import Servant
     )
 import Web.FormUrlEncoded
     ( FromForm (..)
-    , lookupMaybe
-    , parseUnique
-    )
-
-import Cardano.Wallet.Read
-    ( SlotNo (..)
-    , WithOrigin (..)
-    )
-import Cardano.Wallet.UI.Lib.Time.Direction
-    ( Direction (..)
-    )
-import Data.Maybe
-    ( isJust
     )
 
 import qualified Data.ByteString.Lazy as BL
@@ -146,59 +127,60 @@ type Data =
             :> ReqBody '[FormUrlEncoded] TransactionHistoryParams
             :> SessionedHtml Post
 
-instance FromHttpApiData Direction where
-    parseUrlPiece "asc" = Right Asc
-    parseUrlPiece "desc" = Right Desc
-    parseUrlPiece _ = Left "Invalid sorting direction"
 
-data TransactionHistoryParams = TransactionHistoryParams
-    { txHistoryCustomer :: Customer
-    , txHistoryUTC :: Bool
-    , txHistorySlot :: Bool
-    , txHistorySpent :: Bool
-    , txHistoryReceived :: Bool
-    , txHistorySorting :: Direction
-    , txHistoryStartYear :: Int
-    , txHistoryStartMonth :: Int
-    }
 
-instance FromForm Customer where
-    fromForm form = fromIntegral @Int <$> parseUnique "customer" form
 
-instance FromForm TransactionHistoryParams where
-    fromForm form = do
-        utc <- isJust <$> lookupMaybe "utc" form
-        customer <- fromIntegral @Int <$> parseUnique "customer" form
-        slot <- isJust <$> lookupMaybe "slot" form
-        spent <- isJust <$> lookupMaybe "spent" form
-        received <- isJust <$> lookupMaybe "received" form
-        sorting <- parseUnique "sorting" form
-        year <- parseUnique "start-year" form
-        month <- parseUnique "start-month" form
-        pure
-            $ TransactionHistoryParams
-                customer
-                utc
-                slot
-                spent
-                received
-                sorting
-                year
-                month
 
-instance FromHttpApiData SlotNo where
-    parseUrlPiece = fmap SlotNo . parseUrlPiece
 
-instance FromHttpApiData t => FromHttpApiData (WithOrigin t) where
-    parseUrlPiece "Origin" = pure Origin
-    parseUrlPiece t = At <$> parseUrlPiece t
 
-instance ToHttpApiData SlotNo where
-    toUrlPiece (SlotNo t) = toUrlPiece t
 
-instance ToHttpApiData t => ToHttpApiData (WithOrigin t) where
-    toUrlPiece Origin = "Origin"
-    toUrlPiece (At t) = toUrlPiece t
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 type Home = SessionedHtml Get
